Introduced BaseConnectorAlgorithm to encapsulate code common to Mapping and Overlap ConnectorAlgorithms.  Further consolidation is possible.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@258 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/SAMRAI/hier/MultiblockBoxTree.C
@@ -197,7 +197,7 @@ void MultiblockBoxTree::generateTree(
             std::pair<BlockId, BoxTree>(
                blocki->first,
                BoxTree(grid_geometry->getDim(), blocki->second,
-                  blocki->first, min_number))));
+                  blocki->first, static_cast<int>(min_number)))));
       TBOX_ASSERT(insert_return_value.second);
    }
 }
